refactor: 수익률 계산 변경, 구매 로또 출력방식 변경
double 형변환 안해서 생긴 문제 수정
로또를 만들고 정렬하는 게 아니라, view에 넘겨줄 데이터만 따로 정렬해서 넘기도록 수정

+++ src/main/java/lotto/domain/ResultMaker.java
@@ -35,7 +35,7 @@ public class ResultMaker {
     public String calculateRateOfReturn() {
         int investment = bundle.showInvestment();
         calculateTotalPrize();
-        double rateOfReturn = totalPrize / investment * 100;
+        double rateOfReturn = (double) totalPrize / investment * 100;
         return String.format("%.1f", rateOfReturn);
     }
